temporary commit to merge master in
refactor to alarms.py
