+++ tests/test_coordinator.py
@@ -8,7 +8,8 @@ from unittest.mock import patch, Mock
 import pytest
 
 from vent.common import values
-from vent.common.message import ControlSetting, SensorValues, SensorValueNew, Alarm, AlarmSeverity
+from vent.common.message import ControlSetting, SensorValues, SensorValueNew
+from vent.alarms import AlarmSeverity, Alarm
 from vent.common.values import ValueName
 from vent.controller.control_module import ControlModuleBase
 from vent.coordinator import rpc
