+++ vent/gui/main.py
@@ -6,7 +6,8 @@ import os
 
 from PySide2 import QtWidgets, QtCore, QtGui
 
-from vent.common.message import ControlSetting, Alarm, AlarmSeverity
+from vent.common.message import ControlSetting
+from vent.alarms import AlarmSeverity, Alarm
 from vent.common.values import ValueName
 from vent import gui
 from vent.gui import widgets, set_gui_instance, get_gui_instance, styles, PLOTS
@@ -91,7 +92,7 @@ class Vent_Gui(QtWidgets.QMainWindow):
         """
 
         if get_gui_instance() is not None and gui.limit_gui():
-            raise Exception('Instance of gui already running!')
+            raise Exception('Instance of gui already running!') # pragma: no cover
         else:
             set_gui_instance(self)
 
@@ -152,7 +153,6 @@ class Vent_Gui(QtWidgets.QMainWindow):
 
         for control_name, control_params in self.CONTROL.items():
             self.set_value(control_params.default, control_name)
-
 
     def set_value(self, new_value, value_name=None):
         """
